+++ fugitive-darkness-common/src/main/java/com/zer0s2m/fugitivedarkness/common/dto/ContainerGitRepoInstall.java
@@ -2,9 +2,7 @@ package com.zer0s2m.fugitivedarkness.common.dto;
 
 public record ContainerGitRepoInstall(
 
-        String remote,
-
-        String group
+        String remote
 
 ) {
 }